requirements.txt
bottle>=0.12.25
cheroot>=10.0.0
pyScss>=1.4.0
python-decouple>=3.8
